Don't store line numbers in ExecutionContext
Line numbers can be obtained from the instructions, based on the current
instruction index. This doesn't reduce memory usage due to padding, but
it simplifies some of the instruction handling logic.

Initially I made an attempt to change the setup for line numbers
entirely: instead of storing absolute lines per instruction, I
implemented a line offset table similar to the one used in Python.
This ended up reducing the size of bytecode files by about 5%, at the
cost of increasing memory usage by about 5%. Due to the added complexity
of said setup, I decided to not make use of it at this time.