Replace bytecode files with a bytecode image
In this commit we introduce the concept of a bytecode image: a
collection of all the bytecode modules that belong to a program. These
images are parsed at startup, using multiple threads to speed up the
parsing process. We also drop support for loading modules at runtime,
specifying bytecode directories to search, and all functionality related
to runtime module loading.

With these changes, compiled code is reduced to a single bytecode file.
The format of this file is such that it can be easily parsed in
parallel, reducing the startup time. For example, the bytecode image for
Inko's standard library tests (and all dependencies) only takes about
8-9 milliseconds to parse.

These changes remove the need for and performance impact of runtime
module loading. This also reduces the startup time of the VM. In case of
Inko's test suite this reduction is about 20 milliseconds. These changes
also make distribution of programs easier, as you only need to
distribute a single bytecode file. This should also make it easier to
bootstrap the self-hosting compiler in the future, as we could for
example embed the image in a compiled executable; removing the need for
finding it somewhere on the file system.

This fixes https://gitlab.com/inko-lang/inko/-/issues/43