Revamp poetry install retry logic
We now use a shell script in CI to install poetry dependencies before
running `make docs/publish`. This CI script will retry `poetry install`
up to 3 times, removing the need to restart the entire CI job.

This script is not called by our Makefile, as it's only needed in CI.