Use the polling crate for the network poller
This allows us to remove the custom polling code for the various
platforms, as well as allowing us to clean up some of the code used by
the network poller. Finally, this allows us to terminate the poller
thread properly instead of just letting it linger.

This fixes https://gitlab.com/inko-lang/inko/-/issues/222