Make files VM primitives and refactor closing IO
Files are now VM primitives, with the VM exposing three file primitives:
read-only files, write-only files, and read-write files.

In addition to these changes, we refactor how closing of IO resources
works. Prior to this commit, resources were closed by dropping the
object values that wrap them. This created a problem: if a developer
tried to use the resource after disposing of them, the VM would produce
a rather cryptic error. This error would be the result of trying to use
an ObjectValue as e.g. a File, when dropping it previously set it to a
None.

In this commit, the change the VM to provide better support for manually
closing resources. Instead of dropping object values, we use
ObjectValue::close() to close a resource. Closing a resource doesn't
drop it, thus ensuring we don't invalidate the entire ObjectValue.

For files and sockets we just use close() and the Windows equivalents.
After closing such an IO resource, we swap the file descriptor with an
invalid one; preventing accidental closing of reused file descriptors.
When a file or socket is dropped, we close it manually.

For FFI libraries we have to take a slightly different approach, as
different platforms use different approaches to closing them. To close
these types, a Library is now defined as follows:

    #[derive(Clone)]
    pub struct Library {
        inner: Option<ArcWithoutWeak<libloading::Library>>,
    }

The outer Library is just a thing Clone type, kind of like a pointer.
When closing a Library, we simply swap the "inner" field for a None and
drop the ArcWithoutWeak, reducing its reference count. This removes the
need for platform specific logic, and allows efficient copying of
Library types between processes; without running into any race
conditions.