Replace initializer methods with constructors
Object instances can now be created using a special constructor syntax.
This syntax is as follows:

    TypeName { [@attr = value,]* }

For example:

    object Person {
      @name: String
    }

    Person { @name = 'Alice' }

The compiler ensures that all attributes are assigned when using this
syntax. This is more reliable and easier to implement compared to
checking this in an `init` method. This also allows for cleaner APIs.
For example, std::net::TcpStream.from_socket is no longer necessary, as
TcpListener.accept can instead use the constructor syntax.

The new setup also makes it easier to construct an object using rest
arguments. The old `init` method didn't support this, requiring users
to:

1. Define their own `new` method that takes a rest argument.
2. Define an `init` method that takes these arguments as a regular
   array.

In the new setup, one just defines a static method with a rest argument
and that's it.

With the introduction of this syntax, support for initialiser methods
such as the `init` method is removed. In addition, we no longer generate
a `static new` method; instead requiring users to define these
themselves. Not generating these methods gives developers greater
control, simplifies the compiler, and gives developers the option to
define an object that doesn't expose a `new` method.