Add support for writing iterators using generators
Generators are semicoroutines, used to make writing iterators easier. A
generator is created by adding an annotation to a method. When the
annotation is present, whatever expression the method returns is ignored
and instead it returns a generator:

    def numbers => Integer {

    }

Here `=> Integer` signals that the method is a generator yielding values
of type `Integer`. Values are yielded using the `yield` keyword:

    def numbers => Integer {
      yield 10
      yield 20
    }

Generator methods can't return values using the `return` keyword, but
they can still early return without a value:

    def numbers => Integer {
      yield 10
      return
      yield 20
    }

Once a generator returns, it can't be resumed. Generators are
started/resumed using the method `Generator.resume`:

    def numbers => Integer {
      yield 10
      yield 20
    }

    let nums = numbers

    nums.resume # => 10
    nums.resume # => 20

If the generator method throws, the `resume` method will re-throw the
error. In this case, the use of `try` is required:

    def numbers !! String => Integer {
      yield 10
      throw 'oops'
      yield 20
    }

    let nums = numbers

    try nums.resume else 0 # => 10
    try nums.resume else 0 # => 0

To enable this, the Generator type is defined as `Generator!(T, E)`,
with E being the error type. If the generator method doesn't throw, E is
inferred as the Never type. If the throw type of a method is Never, the
use of `try` is not needed.

The Iterator type definition is changed in the same way, and its methods
have had their signature adjusted. Due to a lack of support for default
type parameter values, this makes manually implementing iterators a bit
more annoying. However, thanks to generators this should rarely (if
ever) be necessary, so it's trade-off we're willing to make.

In addition, iterators have a slightly different setup. Instead of
providing the methods `next?` and `next`, iterators now provide
`advance` and `current`. The `advance` method advances the iterator,
returning `True` if a value is produced. The `current` method obtains
this value, returning `Nil` if there isn't any. This new setup makes it
easier to turn generators into iterators, and also makes it a little
easier to write iterators manually.