Replace optional typing with an Option type
An optional type `T` is a type that can either be `T` or `Nil`. These
types were used whenever a value might be left out. For example, getting
a value from an Array using `Array.get` would produce an optional value.

Optional types have the benefit of not introducing any runtime overhead
when a value is present, as no additional objects need to be allocated.
But this also brings several drawbacks, such as:

1. Not being able to distinguish between a Nil being returned explicitly
   versus a Nil being returned due to a lack of a value.
2. Requiring non-trivial type inference code to handle inferring
   optional types properly.
3. Complicating iterator implementations by requiring two methods: one
   to advance the iterator, and one to obtain its value.
4. Soundness problems, such as the one described in
   https://gitlab.com/inko-lang/inko/-/issues/230.
5. Mapping optional types requires special operators or a more complex
   type system, as optional types don't exist at runtime and thus in
   Inko we can't define methods for them.

This commit removes optional typing entirely, replacing it with an
Option type. This solves all the above problems, at the cost of some
additional runtime overhead for allocating Option types. That cost is
something we may be able to optimise away in the future.

As part of this commit, Nil's use case has been reduced significantly.
While still present, it's now used for what it should be used for:
signalling the absence of a value, and that's it. For example, methods
without a return type still return Nil. Similarly, when using the FFI
layer to call a function that returns `void`, a Nil is produced.

This change also means sending messages to Nil is less useful. As a
result, Nil no longer supports messages such as `to_integer`,
`to_string`, etc.

This fixes https://gitlab.com/inko-lang/inko/-/issues/230