Make ByteArray panic for out of bounds operations
Various ByteArray methods, such as ByteArray#[], now panic when given an
out of bounds index. This is a safer default compared to returning
optional types, and makes the index behaviour more consistent with other
(stricter) languages.

In case an optional type is necessary, ByteArray#get can be used which
returns an ?T.