Optimise bounds handling using errors
VM primitives such as array_at, env_get and byte_array_at now throw when
an index or key is out of bounds. This removes the need for additional
bounds checking in the runtime, and makes it easier for the runtime to
decide what default value (if any) to use; instead of it being forced to
use Nil.