Insert a runtime Nil check for the ! operator
This ensures a Nil value can't be produced by accident. This requires
that we fix a few places of the standard library that used this operator
on essentially a nested optional type, otherwise this code will panic.

In addition, when using the ! operator on a ?T where T is a type
parameter, we produce a warning. This ensures that a type parameter set
to Nil won't cause an unexpected runtime panic. In this case an explicit
type cast is necessary.