Add support for template strings
Template strings use backticks, and allow embedding of expressions and
escape sequences. This removes the need for concatenating strings by
hand, or using StringBuffer (at least for most cases where it was used).
For example, instead of this:

    StringBuffer.new('The number is ', number.to_string).to_string

You can now write this:

    `The number is {number}`

Expressions in a template string start with {, and end with a matching
}. If you want the literal {number}, you can write \{number}. That is,
this string:

    `Hello \{name}`

Is the same as this single-quoted string:

    'Hello {name}'

Expressions embedded in a template string must implement the trait
std::conversion::ToString, or the compiler produces a type error.

This fixes https://gitlab.com/inko-lang/inko/-/issues/217