Perform garbage collection in process threads
Instead of using a separate GC coordinator thread, garbage collection is
performed directly in process threads. This reduces collection times as
data doesn't need to be moved across threads and queues, while also
simplifying the setup.

When a collection runs, we also reduce by a custom value, to make up for
the time it may take to perform a collection. This value is arbitrarily
chosen, and may be adjusted in the future.