Don't panic when copying objects
When copying objects that can't be sent across processes, a panic would
be triggered in CopyObject::copy_object(). This commit changes this
method to return a Rust Result instead.

In addition, this commit adds support for sending files across
processes. When a file is copied, the old and new descriptor point to
the same underlying file object. This means operations applied to one
descriptor also affect the other one.