Remove custom implementations of Mirrored
This changes Inko's mirroring setup so the standard library no longer
implements std::mirror::Mirrored for various standard library types.
Instead, we reintroduce std::mirror.reflect. This method uses pattern
matching to determine what mirror to use for an object. It also supports
objects that implement the Mirrored trait.

To facilitate this, you can now write pattern matching expressions like
so:

    match(x) {
      as Array -> { ... }
      else -> { ... }
    }

When the type in an `as ...` pattern is generic, type arguments can't be
specified. In the body to run, the type will be an Array without any
type parameters set. This makes it easier to determine the base type of
an object, which is necessary to make std::mirror.reflect work.

This setup offers several benefits:

1. Importing `std::mirror` no longer pollutes objects with a `mirror`
   method, while still supporting custom mirrors.
2. std::mirror no longer needs to patch Object, bringing us a step
   closer to turning it into a trait.
3. It works better with optional types, as the compiler won't produce an
   error for different implementations of the mirror method for a T and
   Nil in a `?T`.