Reorganise Hash, Equal, and Object
Hash and Equal now provide their methods as default methods, with sane
default behaviour. This makes it easier for the compiler to
automatically implement these traits in the future.

In addition, we add `Object.truthy?` and change the various "if" methods
to make use of this new method. This simplifies their implementations,
and removes the need for some additional closure allocations.

Finally, `Object.equal?` is renamed to `Object.same_object?`, making it
more clear what the purpose of the method is.