Remove support for mutable block arguments
This was used in exactly four places in the standard library, and in all
cases we can get rid of it easily. Mutable arguments can also lead to
confusing errors. If a method re-assigns it without this being obvious,
and a later piece of code in the method expects the initial value to
still be present.