Remove std::marker and std::unknown_message
Instead of using these modules to allow passing Nil to optional types,
the compiler simply has direct knowledge of this being fine. This comes
with the added benefit of not having to pack arguments into an array
when an optional is Nil, making this more efficient.

This fixes https://gitlab.com/inko-lang/inko/-/issues/196.