Allow importing of methods from modules
This makes it possible to import specific methods from a module, instead
of being forced to import the module itself. For example:

    import std::stdio::stdout::(print)

    print('hello')

The use of parentheses for the import is necessary, otherwise the
compiler will try to import the "print" module in the "stdout" module.

While in some cases readability may be improved by importing the module
as a whole (e.g. `ip.parse` is more clear than `parse`), in other cases
leaving out the module _does_ improve readability. Consider for a moment
we have a module called std::loop, which provides a "loop" method.
Importing specific methods allows us to write the following:

    import std::loop::(loop)

    loop { ... }

Instead of having to write this:

    import std::loop

    loop.loop { ... }

This fixes https://gitlab.com/inko-lang/inko/-/issues/158