Use std::loop for creating loops, instead of Block
The Block type provided a few methods: "call", "while_true",
"while_false", and "loop". The "call" implementation didn't take any
arguments, and sending "call" to a block would be optimised into a
RunBlock instruction.

The implementation of the loop methods as it was is unsound. The type
system would allow you to send "while_true" and other messages to all
blocks, even those providing arguments or throwing an error. This means
code such as this would be considered valid at compile-time, but error
at runtime:

    do (number) { number }.while_true { ... }

There is no way we can solve this, at least not without making the type
system overly complicated. Instead, we take a simpler approach: all loop
methods are module methods, defined the "std::loop" module. This
approach gives us several benefits:

1. We don't need to change the compiler.
2. The code is sound.
3. We don't add methods for all blocks. For example, "while_true" for a
   method makes no sense.
4. Prefix notation for loops (`while(condition) { body }`) reads nicer
   than postfix notation (`{ condition }.while_true { body }`).

As part of these changes, "while_false" is removed. This method was
rarely used, and one can just use while() combined with an inverted
condition.

The methods in std::loop aren't added to the prelude yet, requiring a
manual import until this is taken care of; we'll do so in a future
commit.