Refactor how the prelude is set up
The prelude used to be set up by importing std::prelude, and exposing
some of its types to the modules that imported std::prelude. This
approach only worked for exposing constants, not methods. In addition,
it resulted in some standard library modules having access to the
prelude, while others didn't, based on what order they were loaded in.

This commit introduces a new setup. The module std::prelude is renamed
to std::init, and is only used to load the core modules necessary for
every Inko program. The prelude in turn is a small collection of imports
that the compiler injects itself. This allows us to expose a few methods
as part of the prelude, without having to extend the syntax.

The prelude now exposes the following types and methods:

* std::map::Map
* std::option::Option
* std::range::Range
* std::loop.while()
* std::loop.loop()

In addition, the prelude isn't exposed to modules in the std namespace.
This ensures that all standard library modules are treated the same way,
and all require manual importing of the above symbols. While slightly
more annoying compared to not having to import these symbols, it results
in a more consistent development process.