Don't desugar ?T at the parser level
While semantically ?T and Option!(T) are the same, syntactically they
are not. In addition, desugaring ?T at the parser level means a consumer
of the AST can't determine if ?T or Option!(T) was used.

This changes the self-hosting parser so it produces an OptionType AST
node, instead of desugaring ?T into a Constant node.