Fix soundness hole when using type parameters
When using type parameters inside a block, other types were incorrectly
treated as being compatible with the parameter (assuming its
requirements are met). This can lead to soundness issues, such as the
following:

    def push!(A)(values: Array!(A)) {
      values.push(10)
    }

    push(Array.new('foo'))

Here `push` is unsound because the exact type of `A` isn't known until
runtime, meaning it's not safe to push `10` into the array.

To solve this, we introduce the concept of "rigid type parameters".
These are type parameters that can't be further inferred as/assigned new
types, and other types aren't compatible with these type parameters.

The term "rigid type parameter" originates from Haskell. Rigid type
parameters are also called "skolem type parameter", but we don't use
that term because it's unlikely to be widely known or understood.