Infer block return types as Any
When using a closure or lambda in a type signature, leaving out the
return type infers it as Any; not Nil. Inferring the type as Nil can
lead to soundness problems, such as this:

    import std::mirror
    import std::stdio::stdout

    def example(callback: do) -> Nil {
      callback.call
    }

    let result = example { 10 }

    stdout.print(mirror.reflect(result).inspect)

This will print `10` to STDOUT, instead of `Nil`. By inferring
`callback` as `do -> Any` instead of `do -> Nil` we can solve this, as
the above `example` method wouldn't compile (as `Any` isn't compatible
with `Nil`).