Rename the "object" keyword to "class"
Back when Inko was more dynamic/prototype based, the use of `object`
made more sense. As we're planning on changing the memory layout to
resemble a more traditional class based approach, the use of "object"
instead of "class" can be confusing. In addition, users coming from
other languages may wonder why we're not just using "class" instead of
"object".

This commit changes the "object" keyword to "class", solving these
problems.