Refactor std::test and remove panic handlers
The std::test module has been refactored to no longer depend on global
state. Instead, each module must define a `tests` method that takes a
`Tests` type as its argument. This type is used for registering test
groups and tests, without relying on global state (using a process). In
addition, the assertion methods are now located in std::test itself, and
use the naming pattern `assert_X`. They also throw instead of panicking,
requiring the use of `try`.

These changes are made to reduce the places where we panic, to
provide a nicer unit testing interface, and to not rely on global state
(provided through a process).

As part of these changes, support for panic handlers has been removed.
Panics are meant for errors that you can't recover from, but panic
handlers contradict that goal. In addition, they introduce a fair bit of
complexity, and were never truly guaranteed to run. The removal of panic
handlers simplifies the VM, makes it more clear what panics are for,
and allows the VM to be compiled with `panic=abort`.

To handle panics, one should use a system monitoring service of some
sort, such as runit or systemd. Because panics are the result of
software bugs, you can't handle these at runtime. This means the only
sane response is to log it somewhere, and restart the entire program. A
well written program won't panic, so this shouldn't pose a problem in
practise. In addition, external monitoring solutions will always be more
reliable than an in-process monitoring/logging solution, as they aren't
affected by the mayhem caused in the process being monitored.

This fixes https://gitlab.com/inko-lang/inko/-/issues/205.