Don't implement ToArray for Array
Converting an Array to an Array doesn't make much sense, as the input is
already an Array. In addition, users may think `to_array` creates a new
Array when in reality it just returns `self`. Cloning the entire array
and its values would be expensive, and wouldn't bring any benefits.