Support building with Rust 1.50
Rust 1.50 includes some changes related to UNIX file descriptors,
essentially making -1 a reserved value. This means we need to use a
different value to indicate a file that has already been closed. We
don't want to use Option<File> here as this would require double
checking: once to unwrap the option, and once at the OS level to check
if the file descriptor is valid. If/when we move to single ownership
this won't be a problem, as we can just drop the entire File when an
object is dropped.

In addition, we fix some atomic operation deprecations, due to Rust 1.50
deprecating the use of compare_and_swap() in favour of
compare_exchange().