Add note about not accepting all types of changes
In preparation for working on Inko full-time, I'm adding this note. This
should hopefully prevent people from spending a lot of time on changes,
only for me to turn them down. In the long term we may relax the
contribution policy, if we find it won't cost us more effort than it
brings benefits.