New memory management and concurrency support
This commit is the result of 1.5 years worth of work, and changes Inko
dramatically. The garbage collector is replaced using single ownership,
and the Erlang style concurrency API is replaced with one that looks
more like an RPC API, drawing inspiration from Pony. The syntax changed
extensively, the VM has been refactored dramatically, and the compiler
is now written in Rust instead of Ruby.

A lot of other things have changed, far too many to list in this commit.
Instead we'll cover the various changes in greater detail in the release
announcement.