Remove the time dependency
We were using an old version of the time-rs crate as newer versions made
it impossible to get the UTC offset for us, due to soundness issues that
arise when using localtime_r() in combination with Rust's
std::env::set_var.

This commit changes the VM such that we don't use time-rs for getting
the system time and offset, and instead use libc/Windows APIs to do so.
This is made possible by caching environment variables upon startup, and
using this cache for Inko's std::env module. This allows us to remove
the time-rs dependency without introducing soundness holes, reducing the
number of (indirect) dependencies in the process.