Use bit flipping for Float.absolute/opposite
For Float.absolute and Float.opposite we can just change the sign bit
instead of using arithmetic operations. This ensures there's never a
loss of precision.

As part of this I also discovered a compiler bug: constant types were
defined as part of the Expressions pass. If you were to refer to a
constant before its value type is defined, the type would be "unknown".
To fix this, constant types are defined in a separate pass that runs
before the Expressions pass.

This fixes https://gitlab.com/inko-lang/inko/-/issues/239