Fix not exiting scope in OR patterns
If a `match` with an OR pattern was nested in an `if` (and possible
other expressions), the `match` wouldn't exit some of its scopes. If we
then encountered a `return` or `break` in an `else if` or `else`, it
would still see the old scope and attempt to drop its registers. But
since these scopes belong to blocks not a parent of the else if/else,
the compiler would error due to not finding the necessary register
states.

See https://gitlab.com/inko-lang/inko/-/issues/260 for more information.